+++ tests/integrations/test_tornado.py
@@ -11,8 +11,7 @@ class TornadoTests(AsyncHTTPTestCase, IntegrationTest):
 
     def setUp(self):
         super(TornadoTests, self).setUp()
-        bugsnag.configure(use_ssl=False,
-                          endpoint=self.server.address,
+        bugsnag.configure(endpoint=self.server.url,
                           api_key='3874876376238728937',
                           notify_release_stages=['dev'],
                           release_stage='dev',